backend/requirements.txt
# FastAPI Core
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database & Auth
supabase==2.3.0

# Data Processing & ML
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0

# Gradient Boosting Models
xgboost==2.0.3
lightgbm==4.2.0
catboost==1.2.2

# AI/LLM
groq==0.4.1
openai==1.10.0

# Database Connectors
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pymysql==1.1.0
pymongo==4.6.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Explainability
shap==0.44.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# File handling
aiofiles==23.2.1
openpyxl==3.1.2
